requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
httplib2==0.18.1
idna==2.10
logdna==1.5.3
pycparser==2.20
python-decouple==3.4
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
sentry-sdk==0.19.5
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
